adding typvals to all sundials solvers
