+++ Reactions/ReactorCvode.H
@@ -135,8 +135,6 @@ public:
       y_vect, vect_energy, FCunt, dt);
   }
 
-  void SetTypValsODE(const std::vector<amrex::Real>& ExtTypVals) override;
-
 private:
   void checkCvodeOptions() const;
   void allocUserData(
@@ -148,12 +146,10 @@ private:
     amrex::gpuStream_t stream
 #endif
   ) const;
-  void setCvodeTols(void* a_cvode_mem, CVODEUserData* a_udata);
   void freeUserData(CVODEUserData* data_wk);
 
   amrex::Real relTol{1e-6};
   amrex::Real absTol{1e-10};
-  amrex::Array<double, NUM_SPECIES + 1> typVals;
   utils::FlattenOps<Ordering> flatten_ops;
   int m_reactor_type{0};
