+++ Reactions/ReactorNull.cpp
@@ -96,12 +96,6 @@ ReactorNull::react(
 #endif
   return 0;
 }
-
-void
-ReactorNull::SetTypValsODE(const std::vector<amrex::Real>& /*ExtTypVals*/)
-{
-  amrex::Print() << "WARNING: ignoring TypVals for this reactor." << std::endl;
-}
 } // namespace reactions
 } // namespace physics
 } // namespace pele
