+++ Reactions/ReactorRK64.H
@@ -119,8 +119,6 @@ public:
       y_vect, vect_energy, FCunt, dt);
   }
 
-  void SetTypValsODE(const std::vector<amrex::Real>& ExtTypVals) override;
-
 private:
   amrex::Real absTol{1e-10};
   int rk64_nsubsteps_guess{10};
