+++ Reactions/ReactorRK64.cpp
@@ -320,11 +320,6 @@ ReactorRK64::react(
   return (int(avgsteps / amrex::Real(ncells)));
 }
 
-void
-ReactorRK64::SetTypValsODE(const std::vector<amrex::Real>& /*ExtTypVals*/)
-{
-  amrex::Print() << "WARNING: ignoring TypVals for this reactor." << std::endl;
-}
 } // namespace reactions
 } // namespace physics
 } // namespace pele
