+++ Testing/Exec/ReactEval/main.cpp
@@ -432,7 +432,7 @@ main(int argc, char* argv[])
     // -----------------------------------------------------------------------------
     // Set typical values
     // -----------------------------------------------------------------------------
-    if (chem_integrator == "ReactorCvode") {
+    if (chem_integrator == "ReactorCvode" || chem_integrator == "ReactorArkode") {
 #ifdef AMREX_USE_OMP
 #pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
 #endif
